chainfee: fetch new relay fee from bitcoind
This commit adds a function to the BitcoindEstimator that fetches the
currenct minimum relay fee from the bitcoind node and then sets the
BitcoindEstimator minFeePerKW member to this value and also returns the
value. The RelayFeePerKW method is then updated to use this function to
get the latest relay fee from the node. If any error is returned then
the stored relay fee rate is returned.