multi: upate connReqs with new addresses
When a new NodeAnnouncement message is received then check to see if we
have any connection requests active for that node. If we do then ensure
that there are connection requests for all the addresses that were sent
along with the new node announcement and create new connection requests
if needed. This commit fixes the bug asserted in
testReconnectAfterIPChange and so the assertion in the test is changed
too.