peer: add persistent peer manager
This commit adds a PersistentPeerManager which will be used to manager
all persistent peers along with the connection requests for those peers.
The PersistentPeerManager will keep track of all persistent peer
addresses and will subscribe to the graph and update the persistent
peers addresses if a Node Announcement is received that notifies us of
changes to the persistent peer's addresses. If the PersistentPeerManager
receieves address updates for one of its peers then it will also check
if there are any outstanding connection requests for that peer and will
update those accordingly if needed.