server: refactor to use persistent peer manager
This commit is a refactor. It removes the persistentPeers and
persistentConnReqs maps from the server struct and instead uses the new
PersistentPeerManager to register persistent peers and handle their
connection requests. Note that this commit does not call the Start
method of the PersistentPeerManager and so the behaviour does not change
in this commit.