server+refactor: ensure update addresses for persistent peers
In this commit we call the Start method of the PersistentPeerManager to
ensure that it subscribes to the graph and hence keeps track of any
changes to the addresses for its persistent peers. Since the
PersistentPeerManager will now keep track changes in addresses, we no
longer need to check a peer's stored addresses in
peerTerminationWatcher. Instead, on startup when we add the peer as a
persistent peer, we look in the DB and give the PersistentPeerManager
those addresses then. From then on, any updates will come from
NodeAnnouncements which the PersistentPeerManager will keep track of
itself.