lnwallet: fix validateFeeRate
The validateFeeRate function is uses the availableBalance function to
get the current spendable balance of a channel and then adds the old fee
and subtracts the new fee and then checks that this new amount is larger
than the reserve amound. However, this seems to be incorrect since the
availableBalance function already subtracts the reserve balance. This
commit ensures that the corrent base balance is compared with the local
reserve balance.