wtclient: replay unacked updates to task pipeline
In this commit, the bug demonstrated in the previous commit is fixed. So
now if a client wants to remove a tower but the tower has unacked
updates _that are still present in the pending queue_ of the session,
then those will be replayed onto the main task pipeline and the
committed update will be deleted. Note that this only works if the
client is not restarted after the committed updates have been persisted.