lnwire: make MuSig2Nonce TLV type re-usable
Before this commit, any MuSig2Nonce TLV field used within a message is
expected to use the same tlv type number. This is changed in this commit
so that each message must specify which type number it wishes to use.
This is necessary for if there is ever more than one MuSig2Nonce used
within the same message.