lnwire: add Encode and Pack methods for tlv.Records
Sometimes we have a set of records instead of a set of record producers
that we would want to extract from.  So this commit renames the existing
`ExtractRecords` method to `ExtractRecordsFromProducers` (since
extracting from producers is what it is actually doing) and then adds a
new `ExtractRecords` method which extracts from records. The
`ExtractRecordsFromProducers` method calls `ExtractRecords` after first
converting the given `tlv.RecordProducers` into `tlv.Records`.