multi: remove LightningNode from ChannelEdgePolicy
Finally, The LightningNode object is removed frmo ChannelEdgePolicy.
This is a step towards letting ChannelEdgePolicy reflect exaclty the
schema that is on disk.

This is also nice because the `Node` object is not necessarily always
required when the ChannelEdgePolicy is loaded from the DB, so now it
only get's loaded when needed.