wtdb: supply commited update count to PostEvaluateFilterFn
In this commit, we adjust the PostEvaluateFilterFn to also take in a
count representing the number of committed updates (ie, persisted
un-acked updates) that the session has. We also add a new
CountCommittedUpdates option so that the number of committed updates is
not counted unnecessarily. These new options are not yet used.