routing:
We pass the blinded payment associated with an edge to `unifiedEdge`.
This allows us to make some assumptions about the feature set of a blinded
edge without having to be explicity told what those features are.
Specifically, this applies to the `var_onion_option` feature bits which
can be assumed for any node advertising `option_route_blinding`.

See
[this](https://github.com/lightning/bolts/blob/db278ab9b2baa0b30cfe79fb3de39280595938d3/proposals/route-blinding.md?plain=1#L219)
suggestion in the route blinding proposal.

This will allow us to condense the information that a recipient needs to
send to the sender.