build: add handler set implementations
This commit adds a Handler interface which extends the btclog.Handler
interface which is needed for initialising an slog logger.

This commit then also adds an implementation of this interface called
`handlerSets` which basically lets us have a Handler backed by many
Handlers (we need one for the stdout and one for our log file).