autopilot: remove access to graphdb.ChannelGraph pointer
In this commit, we defined a new `autopilot.GraphSource` interface which
describes what the backing graph source for the autopilot agent needs to
implement. Various methods that need the direct graph db pointer are
moved to test files since these methods are only used in tests to
populate the test graph.