multi: remove ChainHash from lnwire.NetAddress
This field is only ever written to and never read from. We remove it so
that later on when we want to convert NetAddress into and from a
protobuf message, we don't need to encode the ChainHash uneccessarily.
This was included in NetAddress when the plan was still to have LND
cater for different blockchains.