graph: create an abstract graphdb.RTx interface
So that we can stop depending on kvdb.RTx directly. The implementation
of this is currently just a kvdb.RTx but eventually this could be a SQL
transaction or even multiple transactions for the case where the query
is being sent to both a local and remote db. The only place where we
want to expose a transaction right now is for pathfinding, so we only do
this for the `graphsession.graph` interface used by pathfinding.