routing: let Graph interface methods take a context.
In preparation for having a GraphSource implementation backed by either
different DB sources (sqlite/postgres which will take a context) or a
remote source over RPC, we update these methods to take a context and
start threading through context to the call sites. The next commit will
aim to remove all the context.TODO()s added in this commit.