refactor: remove graphsession package
In this commit, we remove the graphsession package which added an
uneccessary layer of indirection between the underlying graph and the
interface required by the router. After this commit, the graph "session"
that the router starts, is now fully encapsulated within the
ChannelGraph which means we no longer expose the underlying kvdb.Tx at
all. This simplification/encapsulation will also be important for when
we implement non-kvdb backends for the graph.