discovery: add mockChain to tests and add utxos & blocks
This is in preparation for moving the funding transaction validation
code to the gossiper from the graph.Builder.

See `notifications_test.go` for an idea of what we are trying to emulate
here. Once the funding tx code has moved to the gossiper, then the logic
in `notifications_test.go` will be removed.