discovery: add chan ann creation modifiers
This is copying the behaviour currently used in `graph` tests (see the
graph.edgeCreationModifier) type. So in preparation for moving the
funding tx validation logic to the gossiper, we add that functionality
to the gossiper's tests. It is unused as of this commit.

These will later be used to replace the usage of `setAddEdgeErr` in the
tests.

See `notifications_test.go` for an idea of what we are trying to emulate
here. Once the funding tx code has moved to the gossiper, then the logic
in `notifications_test.go` will be removed.