discovery: add chan ann creation modifiers
This is copying the behaviour currently used in `graph` tests (see the
graph.edgeCreationModifier) type. So in preparation for moving the
funding tx validation logic to the gossiper, we add that functionality
to the gossiper's tests. It is unused as of this commit.

These will later be used to replace the usage of `setAddEdgeErr` in the
tests.