graph: export AddZombieEdge
The `graph.Builder`'s `addZombieEdge` method is currently called during
funding transction validation for the case where the funding tx is not
found. In preparation for moving this code to the gossiper, we export
the method and add it to the ChannelGraphSource interface so that the
gossiper will be able to call it later on.