batch: dont expose kvdb.RwTx in batch.SchedulerOptions
Currently, a few of the graph KVStore methods take the
`batch.SchedularOptions` param. This is only used to set the LazyAdd
option. A SchedularOption is a functional option that takes a
`batch.Request` which has bolt-specific fields in it. This commit
restructures things a bit such that the `batch.Request` type is no
longer part of the `batch.SchedularOptions` - this will make it easier
to implement the graph store with a different DB backend.