multi: add NewTestDB method
In this commit, we add a `test_kvdb.go` file with a single definition of
the `NewTestDB` function. All call-sites of `NewKVStore` are then
replace with calling this new `NewTestDB` function. Later on, we will
add new implementations of this method backed by sqlite and postgres.
When those are added, then build flags will control which version of
`NewTestDB` is called.