htlcswitch: continue threading contexts through
In preparation for `failAliasUpdate` and `createFailureWithUpdate`
calling `FetchLastChannelUpdate` which is a graph DB call that will
eventually take a context.

For `ForwardPackets`, we also make sure to
listen on the newly added context param at any location where we listen
on the `linkQuit` parameter.

This commit adds a few `context.TODO()` calls that will be addressed in
future commits.